requirements.txt
flask-jwt-extended
flask-restx
marshmallow
flask-sqlalchemy
python-dotenv
passlib